+++ src/Objects/CalendarMelanie.php
@@ -37,9 +37,15 @@ use LibMelanie\Config\DefaultConfig;
  */
 class CalendarMelanie extends MagicObject implements IObjectMelanie {
 	/**
+	 * Nom de la table SQL liée à l'objet
+	 * @var string $tableName
+	 */
+	public $tableName;
+	
+	/**
 	 * Constructeur de l'objet, appelé par PDO
 	 */
-	function __construct() {
+	public function __construct() {
 	    // Défini la classe courante
 	    $this->get_class = get_class($this);
 
@@ -57,12 +63,48 @@ class CalendarMelanie extends MagicObject implements IObjectMelanie {
 	}
 
 	/**
+	 * String representation of object
+	 *
+	 * @return string
+	 */
+	public function serialize() {
+		return serialize([
+			'data'        => $this->data,
+			'isExist'     => $this->isExist,
+			'isLoaded'    => $this->isLoaded,
+			'objectType'  => $this->objectType,
+			'primaryKeys' => $this->primaryKeys,
+			'get_class'   => $this->get_class,
+			'tableName'   => $this->tableName,
+		]);
+	}
+
+	/**
+	 * Constructs the object
+	 *
+	 * @param string $serialized
+	 * @return void
+	 */
+	public function unserialize($serialized) {
+		$array = unserialize($serialized);
+		if ($array) {
+			$this->data = $array['data'];
+			$this->isExist = $array['isExist'];
+			$this->isLoaded = $array['isLoaded'];
+			$this->objectType = $array['objectType'];
+			$this->primaryKeys = $array['primaryKeys'];
+			$this->get_class = $array['get_class'];
+			$this->tableName = $array['tableName'];
+		}
+	}
+
+	/**
 	 * Chargement de l'objet
 	 * need: $this->id
 	 * need: $this->user_uid
 	 * @return boolean isExist
 	 */
-	function load() {
+	public function load() {
 		M2Log::Log(M2Log::LEVEL_DEBUG, $this->get_class."->load()");
 		if (!isset($this->id)) return false;
 		if (!isset($this->user_uid)) return false;
@@ -104,7 +146,7 @@ class CalendarMelanie extends MagicObject implements IObjectMelanie {
 	 * Sauvegarde le calendrier
 	 * @return boolean True si c'est une command Insert, False si c'est un Update
 	 */
-	function save () {
+	public function save () {
 		M2Log::Log(M2Log::LEVEL_DEBUG, $this->get_class."->save()");
 		$insert = false;
 		// Si les clés primaires ne sont pas définis, impossible de charger l'objet
@@ -195,7 +237,7 @@ class CalendarMelanie extends MagicObject implements IObjectMelanie {
 	 * (non-PHPdoc)
 	 * @see IObjectMelanie::delete()
 	 */
-	function delete() {
+	public function delete() {
 		M2Log::Log(M2Log::LEVEL_DEBUG, $this->get_class."->exists()");
 		if (!isset($this->tableName)) return false;
 
@@ -247,7 +289,7 @@ class CalendarMelanie extends MagicObject implements IObjectMelanie {
 	 * (non-PHPdoc)
 	 * @see IObjectMelanie::exists()
 	 */
-	function exists() {
+	public function exists() {
 		M2Log::Log(M2Log::LEVEL_DEBUG, $this->get_class."->exists()");
 		// Si les clés primaires et la table ne sont pas définies, impossible de charger l'objet
 		if (!isset($this->tableName)) return false;
@@ -280,7 +322,7 @@ class CalendarMelanie extends MagicObject implements IObjectMelanie {
 	 * L'appel externe n'est donc pas nécessaire (mais cette méthode doit rester public)
 	 * @param bool $isExist si l'objet existe
 	 */
-	function pdoConstruct($isExist) {
+	public function pdoConstruct($isExist) {
 		M2Log::Log(M2Log::LEVEL_DEBUG, $this->get_class."->pdoConstruct($isExist)");
 		$this->initializeHasChanged();
 		$this->isExist = $isExist;
@@ -291,7 +333,7 @@ class CalendarMelanie extends MagicObject implements IObjectMelanie {
 	 * need: $this->id
 	 * @return boolean
 	 */
-	function getAllEvents() {
+	public function getAllEvents() {
 		M2Log::Log(M2Log::LEVEL_DEBUG, $this->get_class."->getAllEvents()");
 		if (!isset($this->id)) return false;
 
@@ -314,7 +356,7 @@ class CalendarMelanie extends MagicObject implements IObjectMelanie {
 	 * @param boolean $is_freebusy Est-ce que l'on cherche des freebusy
 	 * @return boolean
 	 */
-	function getRangeEvents($event_start = null, $event_end = null, $modified = null, $is_freebusy = false) {
+	public function getRangeEvents($event_start = null, $event_end = null, $modified = null, $is_freebusy = false) {
 		M2Log::Log(M2Log::LEVEL_DEBUG, $this->get_class."->getRangeEvents($event_start, $event_end)");
 		if (!isset($this->id)) return false;
 		// DateTime
@@ -369,7 +411,7 @@ class CalendarMelanie extends MagicObject implements IObjectMelanie {
 	 * need: $this->calendar_id
 	 * @return string
 	 */
-	function getCTag() {
+	public function getCTag() {
 		M2Log::Log(M2Log::LEVEL_DEBUG, $this->get_class."->getCTag()");
 		if (!isset($this->id)) return false;
 
@@ -389,7 +431,7 @@ class CalendarMelanie extends MagicObject implements IObjectMelanie {
 	 * Recupère le timezone par défaut pour le
 	 * need: $this->user_uid
 	 */
-	function getTimezone() {
+	public function getTimezone() {
 		M2Log::Log(M2Log::LEVEL_DEBUG, $this->get_class."->getTimezone()");
 		if (!isset($this->user_uid)) return DefaultConfig::CALENDAR_DEFAULT_TIMEZONE;
 
@@ -440,7 +482,7 @@ class CalendarMelanie extends MagicObject implements IObjectMelanie {
 	 * @param string $action
 	 * @return boolean
 	 */
-	function asRight($action) {
+	public function asRight($action) {
 		M2Log::Log(M2Log::LEVEL_DEBUG, $this->get_class."->asRight($action)");
 		return (DefaultConfig::$PERMS[$action] & $this->perm_calendar) === DefaultConfig::$PERMS[$action];
 	}